ADD: parser tree manipulation functions
